Bugfix: don't use i3-lock wrapper script before suspend
Before this fix the screen would lock but the computer wouldn't suspend
until _after unlocking_. This is because the wrapper script tells
`i3lock` not to fork.

Before suspending we _do_ want `i3lock` to fork, so we use the regular
executable. Suspending the computer should put the screens to sleep too.

Extract a new `i3-sleep` wrapper script to keep things DRY.